Move table selection tests to own file.
